+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### v0.11.2
+Make LogStore nullable and handle potential null access
+
+- Update LogTap.store to be nullable.
+- Add null-safe calls (?.) when accessing LogTap.store in WsLogging.kt, Interceptor.kt, server.kt, and LoggingWebSocket.kt.
+- In server.kt, provide an empty list for `/api/logs` if store is null.
+
 ### v0.11.1
 
 Enable resource shrinking and minification in release build; update Proguard rules